7_27
Updated scheduler. Instead of previously holding tasks in an array structure, the data structures for the scheduler are now based on doubly linked lists. This allows for faster context switching via use of a dedicated ready queue. A wait list holds all tasks not in the ready state and these TCBs are processed at each system tick.