8_26
Reorganized memory module according to standard set by the semaphore module. Fixed a bug with the malloc and calloc functions that could result in two tasks trying to claim the same block at the same time. Fixed a bug with the FindIndex function that might return the improper index. Added a switch to rtos_cfg.